+++ jekyll-kuma-plugins/lib/jekyll/kuma-plugins.rb
@@ -2,3 +2,4 @@
 
 require_relative "kuma-plugins/version"
 require_relative 'kuma-plugins/liquid/tags/test'
+require_relative 'kuma-plugins/liquid/tags/policyschema'
